# internal/kadmin/docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
  kadmin.SwaggerErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      data: {}
      message:
        example: invalid request
        type: string
      msg:
        example: invalid request
        type: string
    type: object
...
  kadmin.SwaggerResponse:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      message:
        example: ok
        type: string
      msg:
        example: ok
        type: string
    type: object
...
info:
  contact: {}
  description: KAdmin 前后端分离接口文档。除登录、刷新令牌和公开配置外，请使用 Bearer Token 调用接口。
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: KAdmin API
  version: "1.0"
paths:
...
  /notification-batch/read:
    delete:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kadmin.SwaggerResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: 清空已读通知
      tags:
      - 站内通知
  /notification-batch/read-all:
    patch:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kadmin.SwaggerResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: 全部标记已读
      tags:
      - 站内通知
  /notifications:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 仅未读
        in: query
        name: unreadOnly
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kadmin.SwaggerResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: 查询站内通知
      tags:
      - 站内通知
    post:
      parameters:
      - description: 标题、内容、跳转链接与类型（info/success/warning）
        in: body
        name: payload
        required: true
        schema:
          type: object
      - description: 幂等键（8-128 位）
        in: header
        name: Idempotency-Key
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kadmin.SwaggerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: 发送站内通知
      tags:
      - 站内通知
  /notifications/{id}:
    delete:
      parameters:
      - description: 通知 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kadmin.SwaggerResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除站内通知
      tags:
      - 站内通知
  /notifications/{id}/read:
    patch:
      parameters:
      - description: 通知 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kadmin.SwaggerResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kadmin.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: 标记通知已读
      tags:
      - 站内通知
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"